Move HistoryBindingVirtual to viewer-integration
