+++ build-system/test-configs/forbidden-terms.js
@@ -377,9 +377,9 @@ const forbiddenTermsGlobal = {
       'extensions/amp-access/0.1/login-dialog.js',
       'extensions/amp-story-education/0.1/amp-story-education.js',
       'extensions/amp-subscriptions/0.1/viewer-subscription-platform.js',
+      'extensions/amp-viewer-integration/0.1/history-binding-virtual.js',
       'src/impression.js',
       'src/service/cid-impl.js',
-      'src/service/history-impl.js',
       'src/service/storage-impl.js',
       'src/ssr-template-helper.js',
       'src/service/viewer-cid-api.js',
