+++ src/service/history-impl.js
@@ -15,7 +15,7 @@
  */
 
 import {Deferred, tryResolve} from '#core/data-structures/promise';
-import {dict, map} from '#core/types/object';
+import {map} from '#core/types/object';
 import {getHistoryState} from '#core/window/history';
 
 import {Services} from '#service';
@@ -845,298 +845,19 @@ export class HistoryBindingNatural_ {
 }
 
 /**
- * Implementation of HistoryBindingInterface that assumes a virtual history that
- * relies on viewer's "pushHistory", "popHistory" and "historyPopped"
- * protocol.
- *
- * Visible for testing.
- *
- * @implements {HistoryBindingInterface}
- */
-export class HistoryBindingVirtual_ {
-  /**
-   * @param {!Window} win
-   * @param {!./viewer-interface.ViewerInterface} viewer
-   */
-  constructor(win, viewer) {
-    /** @const {!Window} */
-    this.win = win;
-
-    /** @private @const {!./viewer-interface.ViewerInterface} */
-    this.viewer_ = viewer;
-
-    /** @private {number} */
-    this.stackIndex_ = 0;
-
-    /** @private {?function(!HistoryStateDef)} */
-    this.onStateUpdated_ = null;
-
-    /** @private {!UnlistenDef} */
-    this.unlistenOnHistoryPopped_ = this.viewer_.onMessage(
-      'historyPopped',
-      (data) => this.onHistoryPopped_(data)
-    );
-  }
-
-  /** @override */
-  replaceStateForTarget(target) {
-    devAssert(target[0] == '#', 'target should start with a #');
-    this.win.location.replace(target);
-  }
-
-  /** @override */
-  cleanup() {
-    this.unlistenOnHistoryPopped_();
-  }
-
-  /** @override */
-  setOnStateUpdated(callback) {
-    this.onStateUpdated_ = callback;
-  }
-
-  /**
-   * Gets the history state from a response. This checks if `maybeHistoryState`
-   * is a history state, and returns it if so, falling back to `fallbackState`
-   * otherwise.
-   * @param {*} maybeHistoryState
-   * @param {!HistoryStateDef} fallbackState
-   * @param {string} debugId
-   * @return {!HistoryStateDef}
-   * @private
-   */
-  toHistoryState_(maybeHistoryState, fallbackState, debugId) {
-    if (this.isHistoryState_(maybeHistoryState)) {
-      return /** @type {!HistoryStateDef} */ (maybeHistoryState);
-    } else {
-      dev().warn(
-        TAG_,
-        'Ignored unexpected "%s" data:',
-        debugId,
-        maybeHistoryState
-      );
-    }
-    return fallbackState;
-  }
-
-  /**
-   * @param {*} maybeHistoryState
-   * @return {boolean}
-   */
-  isHistoryState_(maybeHistoryState) {
-    return !!maybeHistoryState && maybeHistoryState['stackIndex'] !== undefined;
-  }
-
-  /**
-   * `pushHistory`
-   *
-   *   Request:  {'stackIndex': string}
-   *   Response: undefined | {'stackIndex': string}
-   *
-   * @override
-   */
-  push(opt_stateUpdate) {
-    const message = /** @type {!JsonObject} */ ({
-      'stackIndex': this.stackIndex_ + 1,
-      ...(opt_stateUpdate || {}),
-    });
-    const push = 'pushHistory';
-    return this.viewer_
-      .sendMessageAwaitResponse(push, message)
-      .then((response) => {
-        const fallbackState = /** @type {!HistoryStateDef} */ (message);
-        const newState = this.toHistoryState_(response, fallbackState, push);
-        this.updateHistoryState_(newState);
-        return newState;
-      });
-  }
-
-  /**
-   * `popHistory`
-   *
-   *   Request:  {'stackIndex': string}
-   *   Response: undefined | {'stackIndex': string}
-   *
-   * @override
-   */
-  pop(stackIndex) {
-    if (stackIndex > this.stackIndex_) {
-      return this.get();
-    }
-    const message = dict({'stackIndex': this.stackIndex_});
-    const pop = 'popHistory';
-    return this.viewer_
-      .sendMessageAwaitResponse(pop, message)
-      .then((response) => {
-        const fallbackState = /** @type {!HistoryStateDef} */ (
-          dict({
-            'stackIndex': this.stackIndex_ - 1,
-          })
-        );
-        const newState = this.toHistoryState_(response, fallbackState, pop);
-        this.updateHistoryState_(newState);
-        return newState;
-      });
-  }
-
-  /**
-   * `replaceHistory`
-   *
-   *   Request:   {'fragment': string}
-   *   Response:  undefined | {'stackIndex': string}
-   *
-   * @override
-   */
-  replace(opt_stateUpdate) {
-    if (opt_stateUpdate && opt_stateUpdate.url) {
-      if (!this.viewer_.hasCapability('fullReplaceHistory')) {
-        // Full URL replacement requested, but not supported by the viewer.
-        // Don't update, and return the current state.
-        const curState = /** @type {!HistoryStateDef} */ (
-          dict({
-            'stackIndex': this.stackIndex_,
-          })
-        );
-        return Promise.resolve(curState);
-      }
-
-      // replace fragment, only explicit fragment param will be sent.
-      const url = opt_stateUpdate.url.replace(/#.*/, '');
-      opt_stateUpdate.url = url;
-    }
-
-    const message = /** @type {!JsonObject} */ ({
-      'stackIndex': this.stackIndex_,
-      ...(opt_stateUpdate || {}),
-    });
-    const replace = 'replaceHistory';
-    return this.viewer_
-      .sendMessageAwaitResponse(replace, message, /* cancelUnsent */ true)
-      .then((response) => {
-        const fallbackState = /** @type {!HistoryStateDef} */ (message);
-        const newState = this.toHistoryState_(response, fallbackState, replace);
-        this.updateHistoryState_(newState);
-        return newState;
-      });
-  }
-
-  /**
-   * Note: Only returns the current `stackIndex`.
-   * @override
-   */
-  get() {
-    // Not sure why this type coercion is necessary, but CC complains otherwise.
-    return Promise.resolve(
-      /** @type {!HistoryStateDef} */ ({
-        data: undefined,
-        fragment: '',
-        stackIndex: this.stackIndex_,
-        title: '',
-      })
-    );
-  }
-
-  /**
-   * `historyPopped` (from viewer)
-   *
-   *   Request:  {'newStackIndex': number} | {'stackIndex': number}
-   *   Response: undefined
-   *
-   * @param {!JsonObject} data
-   * @private
-   */
-  onHistoryPopped_(data) {
-    if (data['newStackIndex'] !== undefined) {
-      data['stackIndex'] = data['newStackIndex'];
-    }
-    if (this.isHistoryState_(data)) {
-      this.updateHistoryState_(/** @type {!HistoryStateDef} */ (data));
-    } else {
-      dev().warn(TAG_, 'Ignored unexpected "historyPopped" data:', data);
-    }
-  }
-
-  /**
-   * @param {!HistoryStateDef} state
-   * @private
-   */
-  updateHistoryState_(state) {
-    const {stackIndex} = state;
-    if (this.stackIndex_ != stackIndex) {
-      dev().fine(TAG_, `stackIndex: ${this.stackIndex_} -> ${stackIndex}`);
-      this.stackIndex_ = stackIndex;
-      if (this.onStateUpdated_) {
-        this.onStateUpdated_(state);
-      }
-    }
-  }
-
-  /**
-   * `getFragment`
-   *
-   *   Request:  undefined
-   *   Response: string
-   *
-   * @override
-   */
-  getFragment() {
-    if (!this.viewer_.hasCapability('fragment')) {
-      return Promise.resolve('');
-    }
-    return this.viewer_
-      .sendMessageAwaitResponse(
-        'getFragment',
-        undefined,
-        /* cancelUnsent */ true
-      )
-      .then((data) => {
-        if (!data) {
-          return '';
-        }
-        let hash = dev().assertString(data);
-        /* Strip leading '#'*/
-        if (hash[0] == '#') {
-          hash = hash.substr(1);
-        }
-        return hash;
-      });
-  }
-
-  /**
-   * `replaceHistory`
-   *
-   *   Request:   {'fragment': string}
-   *   Response:  undefined | {'stackIndex': string}
-   *
-   * @override
-   */
-  updateFragment(fragment) {
-    if (!this.viewer_.hasCapability('fragment')) {
-      return Promise.resolve();
-    }
-    return /** @type {!Promise} */ (
-      this.viewer_.sendMessageAwaitResponse(
-        'replaceHistory',
-        dict({'fragment': fragment}),
-        /* cancelUnsent */ true
-      )
-    );
-  }
-}
-
-/**
  * @param {!./ampdoc-impl.AmpDoc} ampdoc
- * @return {!History}
+ * @return {!Promise<!History>}
  * @private
  */
-function createHistory(ampdoc) {
+function getCreateHistory(ampdoc) {
   const viewer = Services.viewerForDoc(ampdoc);
-  let binding;
+  let bindingPromise;
   if (
     viewer.isOvertakeHistory() ||
     getMode(ampdoc.win).test ||
     ampdoc.win.__AMP_TEST_IFRAME
   ) {
-    binding = new HistoryBindingVirtual_(ampdoc.win, viewer);
+    bindingPromise = Services.virtualHistoryBindingForDocOrNull(ampdoc);
   } else {
     // Only one global "natural" binding is allowed since it works with the
     // global history stack.
@@ -1145,14 +866,22 @@ function createHistory(ampdoc) {
       'global-history-binding',
       HistoryBindingNatural_
     );
-    binding = getService(ampdoc.win, 'global-history-binding');
+    bindingPromise = Promise.resolve(
+      getService(ampdoc.win, 'global-history-binding')
+    );
   }
-  return new History(ampdoc, binding);
+  return bindingPromise.then((binding) => {
+    return (ampdoc) => {
+      new History(ampdoc, binding);
+    };
+  });
 }
 
 /**
  * @param {!./ampdoc-impl.AmpDoc} ampdoc
  */
 export function installHistoryServiceForDoc(ampdoc) {
-  registerServiceBuilderForDoc(ampdoc, 'history', createHistory);
+  getCreateHistory(ampdoc).then((history) => {
+    registerServiceBuilderForDoc(ampdoc, 'history', history);
+  });
 }
